+++ src/manage.c
@@ -2459,7 +2459,7 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
   GHashTable *vts_hash_table;
   osp_credential_t *ssh_credential, *smb_credential, *esxi_credential;
   osp_credential_t *snmp_credential;
-  gchar *max_checks, *max_hosts, *source_iface, *hosts_ordering;
+  gchar *max_checks, *max_hosts, *hosts_ordering;
   GHashTable *scanner_options;
   int ret;
   config_t config;
@@ -2583,11 +2583,6 @@ launch_osp_openvas_task (task_t task, target_t target, const char *scan_id,
   max_hosts = task_preference_value (task, "max_hosts");
   g_hash_table_insert (scanner_options, g_strdup ("max_hosts"),
                        max_hosts ? max_hosts : g_strdup (MAX_HOSTS_DEFAULT));
-
-  source_iface = task_preference_value (task, "source_iface");
-  if (source_iface)
-    g_hash_table_insert (scanner_options, g_strdup ("source_iface"),
-                        source_iface);
 
   hosts_ordering = task_hosts_ordering (task);
   if (hosts_ordering)